ci(native): make clippy gate pass — fix ~60 clippy::all + scope the gate
The audit's "cargo clippy fails" root cause: native_build.yml ran
`cargo clippy --workspace --all-targets -- -D warnings`, but the
workspace opts clippy::pedantic in AS WARNINGS (200+ doc-backtick /
cast-precision nits). `-D warnings` promoted all of those to errors, so
clippy could never pass.

Fixes:
* ~60 real clippy::all warnings across libs, unit/integration tests and
  benches: iterators over index loops, abs_diff/div_ceil/is_multiple_of/
  clamp, slice::from_ref, struct-update syntax, repeat_n, canonical
  partial_cmp (ol_routing min-heap — behavior preserved), redundant
  borrow/closure removal, useless vec!/conversion, runtime.enter() for a
  kept JoinHandle, and a vacuous `|| true` test predicate + compile-time
  const asserts.
* Justified #[allow]s only where the lint is wrong for the context:
  GF(256)/Reed-Solomon matrix loops whose index addresses several arrays,
  crypto Result<_, ()> that deliberately leaks no failure reason
  (decryption-oracle resistance), and signature-transcript builders where
  each arg is a distinct bound field.
* Fixed a PRE-EXISTING bench compile error (sphinx_bench missing the
  SphinxPeelOutcome::Cover match arm) that by itself failed --all-targets.

CI gate (native_build.yml):
* clippy now `-D clippy::all -A clippy::pedantic -A clippy::nursery
  -A clippy::cargo` (pedantic stays advisory, not blocking).
* excludes the pyo3 `one_link_native` cdylib from the pure-Rust job (it
  needs maturin/Python; covered by the python_binding job).
* bounded cargo test with timeout-minutes.

Verified locally: cargo clippy --all-targets -> exit 0; cargo fmt --check
clean; lib tests for all 12 changed crates pass; integration tests pass
(the only failures are pre-existing hardware-sensitive wall-clock
constant-time gates, which also fail on pristine master on this machine —
unrelated to these changes).

+++ native/ol_chunk_store/src/store.rs
@@ -173,7 +173,7 @@ impl ChunkStore {
             Vec::with_capacity(chunk_replay.records.len());
         for (file_id, path) in chunk_files {
             let bytes = std::fs::read(&path)?;
-            let mut cursor: u64 = u64::from(ol_wal::FILE_HEADER_LEN);
+            let mut cursor: u64 = ol_wal::FILE_HEADER_LEN;
             while (cursor as usize) < bytes.len() {
                 // Parse WAL framing manually to recover offsets.
                 let pos = cursor as usize;
